+++ src/framework/Session_local.h
@@ -189,7 +189,7 @@ public:
 
 	// we don't want to redraw the loading screen for every single
 	// console print that happens
-	int lastPacifierTime;
+	float lastPacifierTime;
 
 	// this is the information required to be set before ExecuteMapChange() is called,
 	// which can be saved off at any time with the following commands so it can all be played back
